+++ chapter-5/chapter-5.rb
@@ -1,0 +1,71 @@
+puts 
+puts "# = = = = = = = = = = "
+puts "# Chapter 5"
+puts "# File: chapter-5.rb"
+puts "# = = = = = = = = = = "
+puts
+
+puts 'More About Methods:'
+puts 'Part 1:'
+puts 'hello '.+ 'world'
+puts (10.* 9).+ 9
+puts 'pig ' * 5
+# puts 5 * 'pig'
+puts ''
+
+puts 'Part 2:'
+iCantBelieveIMadeAVariableNameThisLongJustToPointToA3 = 3
+puts iCantBelieveIMadeAVariableNameThisLongJustToPointToA3
+# self.puts iCantBelieveIMadeAVariableNameThisLongJustToPointToA3
+puts ''
+
+puts 'Part 3:'
+var1 = 'stop'
+var2 = 'stressed'
+var3 = 'Can you pronounce this sentence backwards?'
+
+puts 'Before reverse: ' + var1
+puts var1.reverse
+puts ''
+
+puts 'Before reverse: ' + var2
+puts var2.reverse
+puts ''
+
+puts 'Before reverse: ' + var3
+puts var3.reverse
+puts ''
+
+puts 'What is your full name?'
+name = gets.chomp
+puts 'Did you know there are ' + name.length.to_s + ' characters in your name, ' + name + '?'
+puts ''
+
+puts 'Part 4:'
+letters = 'aAbBcCdDeE'
+puts letters.upcase
+puts letters.downcase
+puts letters.swapcase
+puts letters.capitalize
+puts ' a'.capitalize
+puts letters
+puts ''
+
+puts 'Part 5:'
+lineWidth = 50
+puts(                'Old Mother Hubbard'.center(lineWidth))
+puts(               'Sat in her cupboard'.center(lineWidth))
+puts(         'Eating her curds an whey,'.center(lineWidth))
+puts(          'When along came a spider'.center(lineWidth))
+puts(         'Which sat down beside her'.center(lineWidth))
+puts('And scared her poor shoe dog away.'.center(lineWidth))
+puts ''
+
+puts 'Part 6:'
+lineWidth = 40
+str = '--> text <--'
+puts str.ljust  lineWidth
+puts str.center lineWidth
+puts str.rjust  lineWidth
+puts str.ljust(lineWidth/2) + str.rjust(lineWidth/2)
+puts ''