+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 - chapter-4.rb [examples from chapter-4]
 - personal-data.rb [todo from chapter-4]
 - favourite-number.rb [todo from chapter-4]
+- compress big-ruby-program
+
 
 ## [0.0.1] (May 8 2021)